backend/requirements.txt
asgiref==3.5.2
dj-database-url==1.0.0
Django==4.1.3
django-cors-headers==3.7.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
flake8==3.8.4
gunicorn==20.1.0
mccabe==0.6.1
#mysqlclient==2.1.1
Pillow==9.2.0
#psycopg2==2.9.5
pycodestyle==2.6.0
pyflakes==2.2.0
PyJWT==2.5.0
pytz==2020.1
sqlparse==0.3.1
tzdata==2022.6
whitenoise==6.2.0
djangorestframework-simplejwt
APScheduler==3.9.1.post1
twilio==7.16.1
schedule==1.1.0
pdfkit==1.0.0
Jinja2==3.1.2
fpdf==1.7.2
fpdf2==2.6.1